@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.3.3 <=4.0.0"
+  sdk: ">=3.4.3 <=4.0.0"
 
 dependencies:
   flutter:
