@@ -1,4 +1,3 @@
-import 'package:ditonton/search/data/models/search_response.dart';
 import 'package:ditonton/watchlist/domain/entities/watchlist.dart';
 import 'package:equatable/equatable.dart';
 
@@ -30,7 +29,7 @@ class TvSeriesDetail extends Equatable {
   Watchlist toWatchlist() => Watchlist(
         id: id,
         posterPath: posterPath,
-        mediaType: MediaType.TV.name,
+        mediaType: 'tv',
       );
 
   @override
